key strokes broken, cannot get them not to crash slicer

+++ src/sfm_local_chanvese/include/lsops3c.h
@@ -5,13 +5,13 @@
 #include "sparse3c.h"
 #include <vector>
 
-void ls_iteration(double *F, float *phi, short *label, long* dims,
-                  LL* Lz, LL* Ln1, LL* Lp1, LL *Ln2, LL *Lp2, 
+void ls_iteration(float *F, float *phi, short *label, long* dims,
+                  LL* Lz, LL* Ln1, LL* Lp1, LL *Ln2, LL *Lp2,
                   LL *Lin2out, LL* Lout2in);
 
-void ls_iteration_ext(double *F, float *phi, short *label, long* dims,
+void ls_iteration_ext(float *F, float *phi, short *label, long* dims,
                   LL* Lz, LL* Ln1, LL* Lp1, LL *Ln2, LL *Lp2,
-                  LL *Lin2out, LL* Lout2in,LL* Lchanged);
+                  LL *Lin2out, LL* Lout2in, LL* Lchanged);
 
 void ls_mask2phi3c(short *mask, float *phi, short *label, long* dims,
                    LL* Lz, LL* Ln1, LL* Ln2, LL* Lp1, LL* Lp2);
